requirements.txt
# Documentation System Dependencies
# Install with: pip install -r docs/requirements.txt

# Core dependencies
pyyaml>=6.0              # Configuration loading
sentence-transformers>=2.2.0  # Embeddings

# Optional dependencies (graceful degradation if missing)
psutil>=5.9.0            # Resource tracking
aiohttp>=3.8.0           # Async HTTP calls
requests>=2.28.0         # HTTP API calls

# Development dependencies (optional)
pytest>=7.0.0            # Testing
